BC/vanilla_bc/requirements.txt
# =============================================================================
# requirements.txt
# CILRS Vanilla BC — Training + CARLA 0.9.15 Closed-loop
# Python 3.8  |  CUDA 11.8
#
# NOTE: torch / torchvision / torchaudio are NOT listed here because
#       they must be installed with the correct CUDA wheel separately.
#       See the "PyTorch (if not already installed)" block below.
#
# NOTE: CARLA Python API is NOT on PyPI.
#       It is distributed as a .egg file inside the CARLA release folder.
#       See the "CARLA egg" section at the bottom of this file.
# =============================================================================

# ---------------------------------------------------------------------------
# Already installed on this machine (listed for reproducibility on other nodes)
# ---------------------------------------------------------------------------
numpy==1.24.4
Pillow==10.2.0
pygame==2.6.1
networkx==3.0
psutil==7.0.0

# ---------------------------------------------------------------------------
# Data preparation & training utilities  (NOT yet installed)
# ---------------------------------------------------------------------------
openpyxl>=3.1.2          # read .xlsx workzone annotation spreadsheet
tqdm>=4.66.0             # progress bars
pandas>=2.0.3            # optional: useful for CSV analysis

# ---------------------------------------------------------------------------
# Image processing  (NOT yet installed)
# ---------------------------------------------------------------------------
opencv-python>=4.8.0     # image reading fallback + closed-loop frame grabbing

# ---------------------------------------------------------------------------
# CARLA closed-loop dependencies  (NOT yet installed)
# ---------------------------------------------------------------------------
shapely>=2.0.1           # 2D geometry (route/polygon overlap checks)
scipy>=1.10.1            # spatial utilities
py-trees>=2.2.3          # behaviour trees (CARLA ScenarioRunner compatible)
xmlschema>=2.4.0         # CARLA OpenSCENARIO XML parsing
tabulate>=0.9.0          # result summary tables
distro>=1.8.0            # OS detection helper (CARLA internals on Linux)

# ---------------------------------------------------------------------------
# Experiment logging  (optional but recommended)
# ---------------------------------------------------------------------------
# tensorboard>=2.14.0    # uncomment to enable TensorBoard logging
# wandb>=0.16.0          # uncomment to enable Weights & Biases logging
